Initialize package structure using PyScaffold
Ran PyScaffold on the "coffee" package to initialize the package structure. This includes creating the necessary directories, files, and setup files for the package. Many of the placeholder files created by PyScaffold have not yet been filled in.

This commit sets up the basic package structure for the "coffee" package, allowing further development to proceed. This commit does not contain any functional changes to the package itself, but it is an important step towards developing the package further.

Signed-off-by: Christopher Laing <christopher@laing.nz>