Reformat everything with black
Black is a Python auto-formatter. Running it with
'black <path-to-repo-root>' will then reformat all the project code.

This is helpful as the existing code has a lot of end-of-line
whitespace, and many IDEs (notably PyCharm, which I'm using)
automatically remove trailing whitespace, causing spurious changes that
have to be cleaned up before committing.